+++ CHANGELOG.md
@@ -2,4 +2,4 @@
 
 ## [0.1.0] - 2021-07-22
 
-- Initial release
+- Initial release with `anycable_rpc_call_count` and `anycable_rpc_call_runtime` metrics in an AnyCable middleware.